Synapse plot: fix type of tag input field
The type was set to numeric before and since a recent change numeric
fields created in the same way are now enforced to be actually numeric.
This caused the tag input to only accept numbers. This is fixed now by
correctly specifying the input type as 'text'.

Fixes catmaid/CATMAID#1745