Graph widget: fix type of custom regex inputs
The type was set to numeric before and since a recent change numeric
fields created in the same way are now enforced to be actually numeric.
This caused the regex inputs to only accept numbers. This is fixed
now by correctly specifying the input type as 'text' for these fields.

See catmaid/CATMAID#1745