3D viewer: fix type of input of Z plane max zoom level
The type was set to numeric before and since a recent change numeric
fields created in the same way are now enforced to be actually numeric.
This caused the zoom level input to only accept numbers. This is fixed
now by correctly specifying the input type as 'text' for this field.

See catmaid/CATMAID#1745