3D viewer: add option to render catalog panels as PNG
The catalog SVG export dialog has now a new field: "Panel format". It
allows to select between SVG (default), PNG (1 x view size), PNG (2 x
view size) and PNG (4 x view size). The PNG options embed a PNG
rendering of the 3D viewer into the generated catalog SVG instead of
rendering a SVG scene. This allows to reduce the file size, export and
parsing time as well as an exact reproduction of the scene viewed in the
3D Viewer. The different PNG export options allow to increase the DPI
for the individual panels by rendering larger 3D Viewer scene and have
the SVG scale them on the fly, which can be useful for printing.