Node filters: allow partial matches for name checks
Node filter rules can have neuron names to test against. So far, the
rule was only considered applicable for a particular neuron if either no
name was provided to the rule or if the name matches exactly. This
changes now so that also neurons match with a name that is longer than
the rule neuron name but still include it. Exact matching is still
possible, but currently not an option in the UI.