Tests: make expected NPM package versions explicit
We need NPM packages for front-end tests and so far we always installed
the latest version of each required package, which broke our tests with
a recent update of the sinon package (5.0.7 works, 5.0.8 doesn't). To
prevent this in the future and adhere to best practices, these
dependencies are now stored in a package.json file as dev dependencies.
The Travis configuration will install the required packages from this
file and local installation has been made easier as well.