Tracing overlay: initialize graphics data for imported virtual node
Virtual nodes are said to be imported when a request to select a
particular node can't be found locally and is provided by the caller.
This new node is then imported to be selectable. It is however not
enough to just create the node. It is additionally required to create
the graphics data. This makes now all newly imported nodes also visibly
active.

While this could be enough to mark catmaid/CATMAID#1752 as fixed, the
active node selection can still sometimes fail due to other reasons that
I have yet to investigate. Errors unrelated errors to this commit have
been observed when traversing some skeletons: "Could not find node…".

See catmaid/CATMAID#1752