Tracing overlay: add virtual nodes if extra nodes for node selection
If a node is to be selected in a particular tracing overlay it might
happen that the extra node is to be loaded into the current view. In
case both the node and its parent are available for the respective
tracing layer and their connecting edge crosses the tracing layer, a
virtual node is created for completeness.