Graph Widget: allow appending of unselected skeletons
The current append() logic of the Graph Widget is too specific. So far
it removed skeletons from the graph if the skeleton was appended in an
unselected state. This situation however only happens in practice in
situations where it is explicitly not wanted: when subscriptions are
in use. Regular skeleton appending and removal is managed through by the
skeleton source manager. It will respect the selection state of
skeletons and won't append them to a widget if they are unselected. In
this regular scenario, unselected skeletons will never make it into the
append function.

This commit changes this behavior of the Graph Widget. Unselected nodes
can now be appended like selected skeletons. This is useful for
subscripts, because the Graph Widget now won't remove existing skeletons
anymore that are part of subscribed widget that was closed by the user.
It also makes the Graph Widget more consistent with other widgets and
more flexible. Regular append/remove functionality of the Graph Widget
is not affected by this.