Reconstruction sampler: re-evaluate interval node if prev. node was selected
During interval creation, it can happen that the current node is not
used to end the current interval, but the previous one (if it means the
error is smaller and still within the error margin). In this situation,
the current node has to be re-evaluated, because depending on the leaf
handling strategy selected, it might end creating a new interval or
being merged in the last one.