Add Neuron Similarity widget
This widget can be used to create new NBLAST configurations and
computing NBLAST scores based on such a configuration for a query and
target set of neurons.

Creating configurations includes finding random neurons and creating a
similarity matrix. The Similarity matrix can be looked at as a surface
graph and in tabular form, when the "[…]" text is clicked in the
respective configuration row.

Skeleton similarities can be computed based on completed configurations,
i.e. they must have a scoring matrix associated. Once computed, the
scoring can be looked at by clicking on the "[…]" text. It opens a
dialog that shows the top ten matches for each query skeleton including
their score. Negative scoring matches are not shown, since this
indicates it is more likely that the similarity is random. Query
skeletons without any positive score match are hidden by default. The
dialog allows to open the underlying scoring matrix in a new dialog and
to export the whole scoring matrix as CSV.