Tracing layer: add layer setting to load tracing data only in smaller window
The layer settings for the Tracing Layer include now options to define a
"tracing window" width and height as well as whether to show and apply
the tracing window. If enabled (default is disabled), tracing data will
only be loaded if it is in this defined window. Constraining what to
load makes loading tracing data much faster in larger data sets or from
remote locations. This might be especially helpful for remote review.

See catmaid/CATMAID#1732