Tracing overlay: ignore padding if tracing window enabled in update check
The update check makes sure to not do any unnecessary updates and only
allows an update if we the user moved out of the available data. For
regular tracing this includes a padding around the view. This padding
is not used if a tracing window mask is applied and the update check
ignores now the padding if the tracing window is active. Otherwise,
moving the view slighlty wouldn't allow a node update with a tracing
window.