3D viewer: make active node radius adaptive to node radius
The "Radius adaptive active node" checkbox controls whether the active
node size should be adaptive to the node radius, if any. This is enabled
by default. If enabled, the size of the active node will be 1.5x the
size of the radius. If no radius is available or the setting is
disabled, the active node will be sized as before (40 * node scaling
factor).