CLI exporter: export placeholder as roots of new skeletons
Placeholder nodes are now exported as completely new skeletons that are
not linked to their original skeleton, unless it is part of the set of
exported skeletons or the --original-placeholder-context flag is
provided. Exported placeholders can't have any parent nodes, because
those parent nodes are not exported. And they shouldn't be part of their
original skeletons, because if multiple disconnected placeholders refer
to the same skeleton, the skeleton wouldn't be valid anymore (because
its nodes are disconnected).