3D viewer: fix initialization of synaptic and tag spheres
Due to a recent change that introduced better line width computation,
both synaptic spheres and tag spheres where not rendered properly
anymore. The reason was the ID field of individual nodes was still used,
even though it has been removed from it and made available by different
means. This also fixes an error when picking spheres with the Shift +
Click.