Reconstruction sampler: fix arbor adjustments for merge leaf handling
When merging is part of the selected leaf segment handling, CATMAID
removes a potentially previously added node that would mark the old end
of an interval. In such a case, the passed in arbor can be adjusted on
the fly as side-effect. Adjusting the child-parent relationships
involving the removed node was not correct and is fixed now to so that
the nodes surrounding the added node will be linked.