Reconstruction sampler: no new nodes if parent or child are part of interval
If a new node has to be createad during interval creation, there was so
far nothing that prevented this from happening. Now, however, no new
nodes will not be created anymore if a new node involves an interpolated
node as either child or parent. Wihout this the the database location
and the front-end location of client and the database wouldn't match,
preventing the creation of a collinear new point between the edges.