Tracing layer: make field of view available as skeleton source
This makes every tracing layer available as skeleton source. It allows
to pull the currently visible skeletons into other widgets. Tracing
layers won't appear as a default source in other widgets, because it can
contain potentially a lot of skeletons. To do this, skeleton source can
now declare to not be listed as a default selection in other widgets.