Reconstrucion sampler: fix wrong state application for new session
If the "New session" button is pressed, all skeleton filters and
associated data is removed, but the rest of the state is kept (e.g.
interval length). Before, a wrong state reference was used to reset the
session to. This is fixed now and the currently visible data is also
reflected in the internal data structures.