Tracing overlay nodes: access coloring skeleton models indirectly
This is needed, because the `BasicSkeletonSource` used for this in
`TracingOverlay` throws away the old internal object that is referenced
in the node prototype. This made the overlay nodes reference old
skeleton model collection objects. Since they were only used in
connection with skeleton source based node coloring node, this shouldn't
have been a problem often. To fix this, the coloring skeleton models are
now accessed through a Map() like interface (in anticipation of changing
the `BasicSkeletonSource` at some point).