Doc: use Python 3.6 in installation instructions
This updates our default installation instruction to use Python 3.6.
Unfortunately, on Ubuntu 14.04 and 16.04, Python 3.6 requires a custom
PPA to be added to the system. There seems to be no way around this
currently though and given that we have a lot of experience with this
package on production systems, this PPA dependency is deemed okay.

Closes catmaid/CATMAID#1769