Connectivity matrix: add connection fraction display
The new checkbox labeled "Fractions" in the "Main" tab makes it now
possible to display connector fraction instead of an absolute link
number in each cell. The number of connections from one source row to a
target column is divided by the number of total postynaptic connections
that are made to the target skeleton (column). This makes columns better
comparable to each other. Currently, only complete pre-post links
contribute to the to the total number of the total link number of the
target skeletons.

The per-column sums at the bottom are calculated using the sum of all
column links divided by the respective target skeleton total. The row
sums are calculated by adding all row links and dividing them by the sum
of all target links.

Closes catmaid/CATMAID#1747