Reconstruction sampler: fix dead lock in interval creation
During interval creation, the sampler will select the closest real node
by default to end the current interval. If both potential nodes are too
far away, a new node can be created. In case the previous node is
selected, a check was missing if the previous node was the interval
start. If this is the case, the node can't be used, because otherwise
the created interval would have a length of zero.