Tracing overlay: allow node IDs to select to be strings or integers
Due to the change of the nodes data structure to Map, element access is
typed. This causes some issues for other parts of CATMAID that never had
to bother about the type. To mitigate this for now, the node selection
code is now more robust and tries to parse the passed in ID as number.
If this succeeds, the number is used for lookup. Otherwise the original
input string.

See catmaid/CATMAID#1779