Fix a treenode_connector table trigger ambiguity
The on_edit_treenode_update_summary_and_edges() update trigger function
is barely called during regular tracing. When importing data however
this happens more often. This revealed a problem with this trigger
function, an ambiguity was found when referring to `id` in the WHERE
clause. This is fixed with this migration.