Drop existing constraints before ensuring the expected ones exist
In migration 42 some constraints are created on class, class_instance
and class_instance_class_instance that should be there, but have not
been created as part of our current set of migrations. Some CATMAID
instances seem to have these constraints nevertheless, probably from
older migrations or manually added. To be able to deal with this, those
existing constraints are dropped if they exist before the new ones are
created.