3D viewer: warn user if active skeleton isn't loaded but asked to focus
If the user presses the "Focus skeleton" button, the view is focused on
at the center of mass of the active skeleton. If the active skeleton
isn't loaded in the 3D Viewer, the user was so far presented with an
error dialog. This error dialog is now replaced with a simple warning.