3D viewer: make volume picking (Shift + Click) optional
By default, volumes are now not pickable anymore. This means when one
Shift + Clicks on them, one isn't taken to the respective location. If
however a volume in/behind the volume is clicked on, it is now selected.
This scenario is far more common than clicking on volumes. If volumes
should be respected for picking, set "Pickable" in the volume section of
the "View settings" tab.