Tracing overlay: update child edition time in client after node deletion
If a node is deleted and has children, its children are updated to be
linked to the deleted nodes parent. In this operation the child node
changes and thus changes its edit timestamp. This needs to be reflected
in the client to properly pass state checks. Without this a state
matching error was shown when deleting a node with a child node and then
trying to delete its automatically selected parent. This is fixed now.