Tracing overlay: disable rather than obliterate nodes after removal
When nodes are deleted their node data structures are technically not
needed anymore. This however doesn't mean that their Node instances can
be obliterated (e.g. prepared for removal), because these nodes are
still part of a cache. This cache pool is a list with a given size and
we don't know the element's position in this pool. Deleting it would
cause an invalid element in this list, that would still be considered
valid from the cache. If the cache is asked for a free element, it might
end up returning this (obliterated) element and tries to reInit() it.
This will fail if a node has destroyed data that is only created in the
constructor.