Reconstruction sampler: limit merge desicions in merge leaf handling mode
In the leaf handling modes "merge" and "merge-or-create", the fragment
is always merged if possible, i.e. if it is not a twi branchinf off
another segment. This commit adds another condition: the size of the
potentially merged in fragment needs to have a ratio of its length
versus the interval length that is not larger than the new "merge limit"
option. This prevents large leaf segments.