Reconstruction sampler: fix interval rendering issue with leaf segments
The last intervals towards the leaf nodes of a domain sometimes wouldn't
change their color from the previous interval. The problem was that an
update of the internal node -> interval mapping also stored the interval
start node in this data structure. This is by definition not expected
and the color mapper used this information. This is fixed now.