Reconstruction sampler: reset from UI after sampler have been opened
In a new session is started or the active skeleton is used as backbone,
the sampler widger is reset. This has been done so far using the current
widget state as input. This is however only reasonable as long as new
sampler was opened. Otherwise the state would include the properties of
the opened sampler, which would then in turn be used for the reset. This
is confusing and therefore the widget is now reset using the current
state of the sampler UI elements, which is what the user would expect.