Selection table: make new tables last focused
Fix a bug that was causing newly created selection tables to not be set
as the last focused table. This is caused by newly created windows being
created already in the focused state, so when the selection init called
the window focus event, the focus handler (which sets last focused) was
not called because the window was already focused.

Instead invoke the table focus handler directly during initialization.