Reconstruction sampler: add button to show ignored leaf fragment info
The dialog shown when the "Uncovered domain parts" button is pressed in
the Interval step of the Sampler, shows a histogram of the lengths of
all leaf fragments that are not covered by intervals but that are part
of the domain. It uses Plotly.js for this, so the histogram is
interactive by default. When histogram bins are clicked, a Treenode
Table is opened with the respective treenodes listed, which makes it
easier to inspect those sites. The dialog also provides an option to
download the list of all measured leaf fragment lengths.