Skeleton shading: fix sampler-interval shading when nodes where added
If a sampler added nodes, the sampler-interval shading would throw an
error. The problem was the CATMAID.Sampling.intervalEdges() function is
in this case allowed to modify the passed arbor (as sort of cache). So
far only the arbor was updated, but not the location dictionary. This
caused an error on location lookup for newly added nodes. This is fixed
now by caching the position dictionary.