Tracing overlay: fix annotation assignment after swapped merge
When merging two skeleton A into skeleton B, skeleton A won't exist
anymore afterwards (it "loses"). The merge dialog provides a button to
swap the current merge order, i.e. to merge skeleton B into A. The merge
dialog also suggests to keep the name of the losing skeleton as an
annotation on the winning skeleton. When swapping the merge order, this
reference is updates as well. However, while the visual representation
was correct, the actual merge used to the unswapped reference
annotation. This means when after a merge order swap skeleton B would be
merged into skeleton A, the reference annotation kept would still be A.
This is fixed now by updating the merge handler. It was always
referencing the original merge dialog, even after a swap (which creates
a new dialog). It now uses the calling dialog.

Thanks to Katharina Eichler, who noticed this bug.