Connector list: add partner skeleton and partner relation filter
These filters allow to constrain the connector list table based on the
(not displayed) parther(s) of a particular link. If a partner skeleton
source is selected, only rows that link to a connector that also links
to at least one of the partner skeletons are shown. If a partner
relation is selected, the relation has to match this link to a partner
skeleton.