3D viewer: make synapse cluster connector shading more robust
The Synapse Clusters shading mode for connectors won't consider nodes to
be synaptic if there is not an actual pre->post connection. Connector
nodes however are still displayed at these locations and need a color.
Without returning a "not computable" color, an error was raised for
treenodes linked to connectors without a full connection.