Neuron similarity: add optional group permissions to point clouds
This makes it possible to select a group the current user is member of
during the creation of new point clouds. If a point cloud is linked to
a group, it is only visible to members of this group. To be visible for
a group, the group needs to have "can_read" permissions on the point
cloud. If no permission at all is set for a point cloud, everyone can
read it. This makes the common case easier.