Connectivity matrix: add button to download auto-connectivity without display
Showing large connectivity matrices can be slow to display. Therefore a
new button ("Auto-connectivity CSV") allows to export a common case
without displaying the result. The auto-connectivity-matrix, i.e. the
connectivity of a set of skeleton IDs with itself, can now be exported
with a single click of a button. The skeleton IDs are taken from the
currently selected skeleton source. I recently needed this functionality
twice, which is why I added it directly to the widget.