Skeleton merging: add options to handle existing samplers
If the winning skeleton in a merge is referenced by a sampler, users now
have three options on how to update this sampler with the merged in
fragment. If the fragment is merged outside of a domain, nothing special
is happening---it is a regular merge. If the merge treenode is in a
sampler domain, there are currently three options, "Branch", "Domain end"
and "New domain": 1. Branch: add the new fragment to the skeleton
without changing domain end nodes or intervals. This is only allowed if
the merge target is not the start or end of an interval. 2. Domain end:
add a new domain end node right where the merged in fragment starts.
This keeps the new fragment isolated from the sampled domain. 3. New
domain: a new domain is created for the merged in fragment. This also
adds the new domain end node from (2)