DOM: fix tab creation with titles having multiple spaces
The IDs for individual tab containers are created using the tab titles.
Valid IDs can't have spaces, so they are removed. So far this happened
only for the first space and not all of them though. This was of course
only problem if there are multiple spaces in the tab title.