Neuron similarity: add point cloud importer
This allows to import many point clouds at the same time, including
images. The UI works similar to the single file importer in the "Point
cloud" tab, but instead of selecting individual files for CSV point
clouds and images, folders are selected. A single transformation file is
used for all imported point clouds.