SWC export: add options to mark a node as soma
The SWC exporter can now optionally mark a node as soma if it either is
tagged as "soma", if it has a radius larger than a defined value or if
it is the root node. All three conditions can be selected, and they will
be applied in the order they are listed above. If selected, a soma
tagged node will always take precedence.

Closes catmaid/CATMAID#1787