Neuron similarity details: improve predictability of 3D viewer display coloring
If a source object also appears in the target object list, it was so far
re-added when a query result was displayed in the 3D viewer. This caused
its color to change, with is unexpected to the user. This is fixed with
this commit. Objects aren't re-added to the 3D viewer if they were
loaded as source objects already.