Neuron similarity widget: add UI for adding transformed skeletons to a query
To select a transformed skeleton as query or target, such a transform
has to be created in a separate Landmark Widget. These will then be
selectable for both query and target. With the creation of the query, a
new point_set table entry will be created for each referenced
transformed skeleton. IDs of those are referenced by the neuron
similarity query table entry.