Skeleton innervations: make API more CATMAID conform
The previous API didn't comply with our API expectations. The API
contains now proper parameters and documentation. A API endpoint has
been added as well. This API now returns a list of volumes that are
intersected by each skeleton. The list of potential skeletons can be
constrained by a list of skeleton IDs, cable length or node count. The
list of potential volumes can be constrained by a volume annotation.