Volumes: add skeleton innervations API
This allows to find the volumes which have bounding boxes that intersect
with any edges of the passed in skeletons. The API allows optional
parameterization with skeleton length and node count (to additionally
filter the input set) as well as an optional volume annotation to
constrain the set of tested volumes.