Volume manager: add support for more accurate client side filtering
This is enabled by default and causes the widget to load all referenced
volumes and skeletons after the initial bounding box test has been done
on the back-end. This data is then used with the regular skeleton filter
infrastructure to find only skeletons that actually intersect with the
volumes and not only their bounding boxes.

Closes catmaid/CATMAID#1621