3D viewer: fix three.js maintaining the camera up vector in animations
If the camera turns up side down during an animation, e.g. when the X
axis is the rotation axis, it used to flip around to maintain its up
vector. This was very confusing visually, which is why this is now
fixed. Thanks to Katharina Eichler for bringing this up!