Update commit representation in version string
The CATMAID version string is still available from the VERSION attribute
of the settings object. This version is made up from two components: a
base version and commit information. The base version is explicitly
stored in the mysite.utils module (where the `get_version()` function
lives). It is updated with every release by the `make-releas.py` script.

The commit ID is retrieved by first trying to execute Git and find the
latest commit in the local directory. If this isn't possible (e.g. no
Git or no repository available), CATMAID tries to open the file
'/home/git-commit', which in turn is generated by the Docker build. In
regular instances, this file shouldn't be available. If this fails as
well, the commit ID will become "unknown".

Also the CATMAID version information is now printed on start up on the
back-end, too.