Tracing layer: make Alt + Click behave more consistently
Sanja Jasek reported that the behavior of Alt + Click is dependent on
whether a treenode or a connector node is selected, which can be pretty
confusing. Alt + Click used to show a link type menu when a treenode was
selected and created a gapjunction_with link when a connector node is
selected. This is now changed so that a link type menu is always shown.
This makes Alt + Click more consistent and versatile, because other
connector link types can now be selected when a connector is already
selected.