Connector API: improve many_to_many and one_to_many API
This makes it easier to add new connector link types and makes the SQL
function faster, because the relation is looked up by ID, which removes
one join and is a faster test.

Also, this makes the actual construction of the SQL query nicer and more
safe by using psycopg2's parameterization.