Tracing layer: add option to only show N most recently edited skeletons
This selects the N most recently edited skeletons in the field of view
using the skeleton summary table. Skeletons selected this way are added
to the skeletons that have been potentially selected with the "N largest
skeletons" filter. This option is accessible from the layer settings in
the stack viewer. Node cache functionality has been updated as well.