Skeleton summary table: store last editor ID
The last editor is now stored for each skeleton. On node deletion, the
user performing the deletion can be passed to the trigger by setting the
transaction variable 'catmaid.user_id' and is stored in the respective
skeleton summary. Skeleton updates, including skeleton ID, will update
the editor accordingly.

This requires an update of the respective trigger functions. Performance
wise this doesn't seem to make a difference for splitting and merging.