Tracing layer: optionally prevent loading of skeletons edited last by a user
This adds an option the tracing layer options to select a user which
will be used to prevent skeleton tracing data from loading if that user
was the one editing a particular skeleton last. The main motivation to
have this is to prevent parts of a data set that many users don't want
to see by default from loading. An example is a large set of
skeletonized segmentation data, which is imported by a dedicated user.
Users can show this data if they want, but in case the quality is
otherwise not en par with the manual reconstructions, users might want
to hide this by default and lower loading times.