Skeleton split/merge: build domain graph without recursion
If reconstruction samplers exist for a skeleton that is split or merged,
the graph representing a particular sampler domain needs to be
constructed to check how a split-off part or merged-in part will affect
the sampler. This was so far done with recursion, but unfortunately for
larger skeletons this hits Python's recursion limit. Therefore, the
recursion is now unwrapped and the domain graph is constructed
incrementally.

This fixes merging and splitting larger skeletons with samplers.