Tracing overlay: allow configuration of tracing data mirrors
Similar to image data mirrors, it is now possible to configure read-only
tracing data mirror servers from which the tracing data will load all
data except for the active node, which will be read from the main
server. This is particularly useful to improve the loading time if
connecting to the main server from a remote location with reasonable
local server infrastructure.

To make this work reliably, it is expected that physical replication is
setup on the database level that mirrors the main server constantly.
Documentation on how to do this will follow. A separate CATMAID instance
needs to be setup on the mirror server as well. The webserver running
this needs to be configured to allow CORS rquests.

To configure mirrors in the CATMAID front-end, the "Read-only CATMAID
mirrors" option in the settings widget can be used together with the
"Read-only mirror index".

A list of read-only CATMAID mirror servers and optionally API keys is
used to query tracing data from. In the setting, individual entries are
separated by commas and have The form "url|apikey. URLs need to include
the protocol, e.g.:

"https://example.com/catmaid/|apikey1, https://example2.com/catmaid2/"