Tracing overlay: warn in front-end about root node deletion with children
A root node with child nodes can't be deleted at the moment. So far the
back-end returned an error. This will still happen if the request is
actually sent, but the front-end catches this situation now preemptively
and warns the user without sending the request. This is much faster than
waiting for the server error response.