Skeleton split/merge: fix wrong sub-graph creation
In 6a0948cb7c3c the sub-graph creation for a split-off part or merged-in
part has been changed from a recursive implementation to an iterative
one. This introduce a bug that would lead to wrong sub-graphs. This is
fixed now and splits/merges of skeletons with samplers should now work
reliably.