Upgrade Django from v2.0.9 to v2.1.3
This update is done to fix a few CI test problems with Django 2.0.9. It
also seems to fix a Python 3.7 relatead problem on MacOs. More details
on this update can be found here:

  https://docs.djangoproject.com/en/2.1/releases/2.1/

I went through the changelog and most of it doesn't apply to our use
case. A few minor changes had to be performed though:

- Replaced "{% load admin_static %}" with "{% load static %}" in
  tempates.
- Replaced "contrib.auth.views.change_password()" use with class based
  version: NonAnonymousPasswordChangeView, which is a non-anonymous user
  version of the PasswordChangeView.
- Fix JSONField default argument to be a callable rather than an
  instance, so it is not shared (performancetests.TestView.data:
  (postgres.E003)).
- Remove the explicit wrap_label reference when creating downsampling
  widget in template. This option is now part of the passed in options.
- Updated tests to reflect new default 'view' permission for models.

See catmaid/CATMAID#1766