Changelog: add needed removal of .pyc migration files
The removal of all .pyc files for all migrations in CATMAID's Django
applications (catmaid/, performancetests/, pgcompat/) is needed due to
the upgrade to Django 2.1. Starting with this Django version migrations
are loaded from .pyc files. In case both Python 2 and Python 3 have been
used before with the same files, all .pyc files need to be removed. This
is mentioned in the Django 2.1 release notes:

https://docs.djangoproject.com/en/2.1/releases/2.1/