Skeleton annotations: manage UI updates through events
It's not the SkeletonAnnotations singleton's business to update the UI
of the tracing tool or the mouse cursor of the tracing overlay. Instead
it now fires an event if the interaction mode changed. The tracing tool
listens to this and updates the UI accordingly and the tracing overlay
updates the cursor in use.