Selection table: allow adding skeletons w/o review information
So far, the Selection Table refused to load skeletons for which it
couldn't find review information for a skeleton. While this shouldn't
happen in the first place (we can assume the summary table has an entry
for each skeleton), the Selection Table should also not depend on this
review information alone. When loading large datasets, it is not
uncommon to first load the data without indexes or summary information
and create this information in a separate step. In situations like it is
useful to inspect skeletons quickly in the 3D Viewer (using the
Selection Table), even if the summary information isn't available yet.