Neuron search: improve selection behavior of expanded annotations
Annotations can be expanded in the Neuron Search widget which leads to
all its annotated entities being displayed in the result list. If the
checkbox next to the expanded annotation is now clicked, the neurons
all expanded neurons will be selected accordingly. This is currently
only enabled for expanded annotations, but it might make sense to update
this in the future to also select all neurons of an annotation if it is
not expanded.