Tracing overlay: avoid repeated test when adding virtual nodes
When adding virtual nodes during tracing data loading, virtual nodes are
created when neither child or parent of an edge are on the current
section, but the edge intersects with it. The test whether this is the
case has been done twice before in the common case (loading tracing
data). This duplicate test is now removed as a micro-optimization.