Tracing overlay nodes: use a shared intersection storage to reduce GC
Getting the child and parent intersections is done very frequently and
for views > 10,000 nodes a single update can cause many minor garbage
collections. In an effort to microoptimize the display of larger FOVs, a
shared variable is now used. Given JavaScript is single-threaded, this
is no problem in terms of concurrency.