Landmark widget: add options to define remote instance as source
By default, the Landmark Widgets shows only controls to source skeletons
and landmark groups from the current project. With the new option
"Source other projects" a slightly different UI is presented. It allows
users to select a remote CATMAID instance (can be configured in Settings
Widget) as well as a project and a source landmark group from it. To
select neurons from the remote project, an annotation can be used.