Authentication: improve robustness for anonymous user variants
Django comes with its own AnonymousUser model, which isn't a real user
and mocks the User model interface only to some extent. It lacks for
instance the fields for first name and last name. CATMAID avoids this
user and replaces references in an incoming request with Guardian's
AnonymousUser, which is a real user. This is done through a middleware
(catmaid.middleware.AnonymousAuthenticationMiddleware). Unfortunately,
this breaks down in some cases where DjangoRestFramework is used,
because it does its own authentication logic and ignores middleware
decisions and falls back to Django's AnonymousUser.

To make CATMAID's authentication decorators work regardless, permission
errors now don't reference first/last name anymore and the user ID is
used explicitly as selector for permission lookups. Without this, DRF
requests (e.g. /settings or /bookmarks) would fail with an error like
this:

  TypeError: int() argument must be a string, a bytes-like object or a number, not 'AnonymousUser'

Surprisingly, this error did only show up in a very small set of CATMAID
instances and it is not entirely clear why other instances didn't result
in the same behavior. There is however currently no real motivation to
look into this more closely, because this happened so rarely and is
fixed now.

While the changes explained above are enough to make this error
disappear, the root of this error is the fact that DRF falls back to
Django's AnonymousUser rather than Guardian's AnonymousUser (which is
an actual User model instance). To fix this, the DRF setting
'UNAUTHENTICATED_USER' is set to 'guardian.utils.get_anonymous_user',
which allows DRF to fall back to Guardian's anonymous user, which
avoids the above problem altogether. Still, the above changes are useful
to improve robustness with respect to future changes.